move fee module to common
