+++ src/transaction/mod.rs
@@ -4,7 +4,6 @@ pub mod declare_v2;
 pub mod deploy;
 pub mod deploy_account;
 pub mod error;
-pub mod fee;
 pub mod invoke_function;
 pub mod l1_handler;
